Refactor StopActionsDropdown responsibility
The responsibilities were a bit scattered around. So moved the onAddToRoute
and onRemoveFromRoute handlers to StopActionsDropdown component, instead
of passing those handlers through two components. Instead we now just pass
the route through the components, which is in my opinion a bit more clear.
Also this decouples StopActionsDropdown and RouteStopsSection.
